project is uploaded
Archive file contains a folder with folders inside with the codes and images in three folders. Folder 'imgs' contains images to be processed. In fact there may be any quantity of images of cards. Final aim is to have list of files with symbols that are appropriate to suites and numbers of cards on that file. 's' stands for spades, 'c' for clubs, 'd' for diamond, 'h' for hearts. Numbers are from 2 to 9, then, as usually, come A, J, Q AND K. So the line following the name of file 'Jc10h7dQs' means that this file has J of clubs, 10 of hearts, 7 of diamonds and Queen of spades.
I used comparison of quantity of match of colors of pixels and has chosen the biggest number of matches to indicate the suite and number. It works.
But, I suppose, to make the image recognition more universal and exact, the arrays should be used. 
What is your thought?